more PR changes. added tests, and moved RegisterMessages.
