+++ cardinal/cardinal.go
@@ -2,10 +2,12 @@ package cardinal
 
 import (
 	"errors"
+	"reflect"
 	"strconv"
 
 	"github.com/rotisserie/eris"
 	"pkg.world.dev/world-engine/cardinal/iterators"
+	"pkg.world.dev/world-engine/cardinal/message"
 	"pkg.world.dev/world-engine/cardinal/storage/redis"
 	"pkg.world.dev/world-engine/cardinal/system"
 	"pkg.world.dev/world-engine/cardinal/types"
@@ -114,6 +116,16 @@ func RegisterMessagesByName(w *World, msgs ...types.Message) error {
 		)
 	}
 	return w.msgManager.RegisterMessagesByName(msgs...)
+}
+
+func RegisterMessage[In any, Out any](world *World, name string, opts ...message.MessageOption[In, Out]) error {
+	msgType := message.NewMessageType[In, Out](name, opts...)
+	err := world.registerMessagesByName(msgType)
+	if err != nil {
+		return eris.Wrap(err, "failed to register message")
+	}
+	typeValueOfMessageType := reflect.TypeOf(*msgType)
+	return world.GetMessageManager().RegisterMessageByType(typeValueOfMessageType, msgType)
 }
 
 func RegisterQuery[Request any, Reply any](
